agent/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.28.1
pydantic==2.10.3
pydantic-settings==2.7.1
python-json-logger==3.2.1
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.0.2
chromadb-client==0.6.3
